+++ demo/api/src/news/entities/news.entity.ts
@@ -1,5 +1,5 @@
 import { BlockDataInterface, RootBlock, RootBlockEntity } from "@comet/blocks-api";
-import { CrudField, CrudGenerator, DamImageBlock, DocumentInterface, RootBlockDataScalar, RootBlockType } from "@comet/cms-api";
+import { CrudField, CrudGenerator, DamImageBlock, DocumentInterface, EntityInfo, RootBlockDataScalar, RootBlockType } from "@comet/cms-api";
 import { BaseEntity, Collection, Embeddable, Embedded, Entity, Enum, OneToMany, OptionalProps, PrimaryKey, Property } from "@mikro-orm/core";
 import { Field, ID, InputType, ObjectType, registerEnumType } from "@nestjs/graphql";
 import { IsString } from "class-validator";
@@ -7,6 +7,12 @@ import { v4 as uuid } from "uuid";
 
 import { NewsContentBlock } from "../blocks/news-content.block";
 import { NewsComment } from "./news-comment.entity";
+
+export enum NewsStatus {
+    Active = "Active",
+    Deleted = "Deleted",
+}
+registerEnumType(NewsStatus, { name: "NewsStatus" });
 
 export enum NewsCategory {
     Events = "Events",
@@ -32,6 +38,7 @@ export class NewsContentScope {
     language: string;
 }
 
+@EntityInfo<News>((news) => ({ name: news.title, secondaryInformation: news.slug }))
 @RootBlockEntity()
 @ObjectType({
     implements: () => [DocumentInterface],
@@ -39,7 +46,7 @@ export class NewsContentScope {
 @Entity()
 @CrudGenerator({ targetDirectory: `${__dirname}/../generated/` })
 export class News extends BaseEntity<News, "id"> implements DocumentInterface {
-    [OptionalProps]?: "createdAt" | "updatedAt" | "category"; // TODO remove "category" once CRUD generator supports enums
+    [OptionalProps]?: "createdAt" | "updatedAt" | "category" | "status"; // TODO remove "category" once CRUD generator supports enums
 
     @PrimaryKey({ type: "uuid" })
     @Field(() => ID)
@@ -57,6 +64,10 @@ export class News extends BaseEntity<News, "id"> implements DocumentInterface {
     @Field()
     title: string;
 
+    @Enum({ items: () => NewsStatus })
+    @Field(() => NewsStatus)
+    status: NewsStatus = NewsStatus.Active;
+
     @Property()
     @Field()
     date: Date;
@@ -64,10 +75,6 @@ export class News extends BaseEntity<News, "id"> implements DocumentInterface {
     @Enum({ items: () => NewsCategory })
     @Field(() => NewsCategory)
     category: NewsCategory = NewsCategory.Awards; // TODO remove default value once CRUD generator supports enums
-
-    @Property()
-    @Field()
-    visible: boolean;
 
     @RootBlock(DamImageBlock)
     @Property({ customType: new RootBlockType(DamImageBlock) })