+++ packages/admin/blocks-admin/src/blocks/common/AdminComponentButton.tsx
@@ -1,6 +1,4 @@
 import { Button } from "@mui/material";
-import makeStyles from "@mui/styles/makeStyles";
-import clsx from "clsx";
 import * as React from "react";
 
 import { AdminComponentPaper } from "./AdminComponentPaper";
@@ -15,22 +13,18 @@ interface Props {
 }
 
 export function AdminComponentButton({ variant, size, ...buttonProps }: Props): React.ReactElement {
-    const classes = useStyles();
-
     return (
         <AdminComponentPaper disablePadding>
             <Button
                 fullWidth
                 color={variant === "primary" ? "primary" : "info"}
-                className={clsx(size === "large" && classes.buttonSizeLarge)}
+                sx={({ spacing }) => ({
+                    ...(size === "large" && {
+                        padding: spacing(4),
+                    }),
+                })}
                 {...buttonProps}
             />
         </AdminComponentPaper>
     );
 }
-
-const useStyles = makeStyles((theme) => ({
-    buttonSizeLarge: {
-        padding: theme.spacing(4),
-    },
-}));