+++ packages/api/cms-api/src/page-tree/page-tree.module.ts
@@ -10,8 +10,10 @@ import { DocumentSubscriberFactory } from "./document-subscriber";
 import { PageTreeNodeBaseCreateInput, PageTreeNodeBaseUpdateInput } from "./dto/page-tree-node.input";
 import { AttachedDocument } from "./entities/attached-document.entity";
 import { PageTreeNodeBase } from "./entities/page-tree-node-base.entity";
-import { defaultReservedPaths, PAGE_TREE_CONFIG, PAGE_TREE_ENTITY, PAGE_TREE_REPOSITORY, SITE_PREVIEW_CONFIG } from "./page-tree.constants";
+import { defaultReservedPaths, PAGE_TREE_CONFIG, PAGE_TREE_ENTITY, PAGE_TREE_REPOSITORY } from "./page-tree.constants";
 import { PageTreeService } from "./page-tree.service";
+import { PageTreeNodeDocumentEntityInfoService } from "./page-tree-node-document-entity-info.service";
+import { PageTreeNodeDocumentEntityScopeService } from "./page-tree-node-document-entity-scope.service";
 import { PageTreeReadApiService } from "./page-tree-read-api.service";
 import { SitePreviewResolver } from "./site-preview.resolver";
 import type { PageTreeNodeInterface, ScopeInterface } from "./types";
@@ -28,7 +30,6 @@ interface PageTreeModuleOptions {
     Documents: Type<DocumentInterface>[];
     Scope?: Type<ScopeInterface>;
     reservedPaths?: string[];
-    sitePreviewSecret: string;
 }
 
 @Global()
@@ -86,15 +87,11 @@ export class PageTreeModule {
                     inject: [PageTreeService],
                 },
                 documentSubscriber,
-                {
-                    provide: SITE_PREVIEW_CONFIG,
-                    useValue: {
-                        secret: options.sitePreviewSecret,
-                    },
-                },
                 SitePreviewResolver,
+                PageTreeNodeDocumentEntityInfoService,
+                PageTreeNodeDocumentEntityScopeService,
             ],
-            exports: [PageTreeService, PageTreeReadApiService, AttachedDocumentLoaderService],
+            exports: [PageTreeService, PageTreeReadApiService, AttachedDocumentLoaderService, PageTreeNodeDocumentEntityScopeService],
         };
     }
 }