+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/admin-theme
 
+## 6.4.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.4.0
+
+## 6.3.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.3.0
+
 ## 6.2.1
 
 ### Patch Changes